SuffixTree: Show problem with multiple calls to insert_ukkonen().
When we follow a suffix link, we don't change the active_edge (an index
of the inserted string, to identify an edge from the active_node) and
active_length (the position along the edge). So, if that path was
actually created by a previous construction, it might be split into
several edges. So, even if we allow for that, the active_length alone is
not enough to specify the position along that path.