test: suffix_tree: Keep the input string alive
This is not pretty, but it's what the SuffixTree class requires,
to be as efficient as possible, avoiding copying. It uses only
indexes into the strings.

A truly general-purpose API would not require this, though would
maybe still allow it.

This fixes the tests with g++. It was just luck that it worked
with clang.